finishing implementing home flow
