+++ .idea/misc.xml
@@ -32,7 +32,7 @@
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/fragment_sending_data.xml" value="0.1675929884566054" />
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/fragment_services.xml" value="0.21" />
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/fragment_telephone.xml" value="0.1675929884566054" />
-        <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/fragment_transfer_detail.xml" value="0.2464065708418891" />
+        <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/fragment_transfer_detail.xml" value="0.5" />
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/fragment_waiting_transaction.xml" value="0.2" />
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/item_cardview_contact.xml" value="0.33" />
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/layout/item_cardview_transactions.xml" value="0.33" />
@@ -41,10 +41,14 @@
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/menu/bottom_nav_menu.xml" value="0.1675929884566054" />
         <entry key="..\:/Users/pablo/AndroidStudioProjects/bankodemia/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.195" />
         <entry key="app/src/main/res/drawable/bg_skeleton_component.xml" value="0.1845" />
-        <entry key="app/src/main/res/layout/activity_home2.xml" value="0.1" />
+        <entry key="app/src/main/res/drawable/ic_logomini.xml" value="0.1285" />
+        <entry key="app/src/main/res/drawable/ic_services.xml" value="0.1285" />
+        <entry key="app/src/main/res/drawable/nav_view_item_textcolot.xml" value="0.1285" />
+        <entry key="app/src/main/res/layout/activity_home2.xml" value="0.2715057373046875" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.14945652173913043" />
-        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.13793103448275862" />
-        <entry key="app/src/main/res/layout/item_cardview_transactions.xml" value="0.47972529970366384" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/item_cardview_transactions.xml" value="0.2934936523437501" />
+        <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.1" />
       </map>
     </option>
   </component>
