+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.example.bankodemia">
 
     <uses-permission android:name="android.permission.INTERNET"/>
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -14,7 +15,7 @@
         <activity android:name=".ui.view.MainActivity"
             android:exported="false"/>
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.view.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
