+++ app/src/main/java/com/example/bankodemia/network/api/UserAPI.kt
@@ -8,5 +8,5 @@ import retrofit2.http.Header
 interface UserAPI {
     @GET("/users/me/profile")
     // TODO- remove harcoded token when logic is available
-    suspend fun getUserProfileInfo(@Header("Authorization") token: String = "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2MjJiYjEzNDhjZTZjNDc4ZDBlMWJmZTYiLCJpYXQiOjE2NDczMTE0MDYsImV4cCI6MTY0NzMxNTAwNn0.RB5gr-_yFKCrtyyARUjPSoU4j7BIASc5HMFkUd8rPMo"): Response<User.UserProfile>
+    suspend fun getUserProfileInfo(@Header("Authorization") token: String = "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2MjJiYjEzNDhjZTZjNDc4ZDBlMWJmZTYiLCJpYXQiOjE2NDczOTgzMTIsImV4cCI6MTY0NzQwMTkxMn0.BNGiVOs0WdrGdps8DLXYja_h0dhMbS7Q5Xxu3u8t9Ng"): Response<User.UserProfile>
 }
