+++ app/src/main/java/com/example/bankodemia/ui/view/AddContactFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bankodemia.UI.View
+package com.example.bankodemia.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -10,12 +10,10 @@ import com.example.bankodemia.core.types.FieldTypeEnum
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.findNavController
 import com.example.bankodemia.R
-import com.example.bankodemia.UI.viewModel.AddContactViewModel
-import com.example.bankodemia.UI.viewModel.CardsViewModel
+import com.example.bankodemia.ui.viewModel.AddContactViewModel
 import com.example.bankodemia.core.activateButton
 import com.example.bankodemia.core.validateField
 import com.example.bankodemia.databinding.FragmentAddContactBinding
-import com.example.bankodemia.ui.view.Fields
 
 class AddContactFragment : Fragment(), Fields {
     private var _binding: FragmentAddContactBinding? = null
