+++ app/src/main/java/com/example/bankodemia/ui/view/CardsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bankodemia.UI.View
+package com.example.bankodemia.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,15 +6,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
-import com.example.bankodemia.R
-import com.example.bankodemia.UI.viewModel.CardsViewModel
-import com.example.bankodemia.UI.viewModel.ServicesViewModel
+import com.example.bankodemia.ui.viewModel.CardsViewModel
 import com.example.bankodemia.databinding.FragmentCardsBinding
-import com.example.bankodemia.databinding.FragmentServicesBinding
 
+class CardsFragment : Fragment() {
 
-class ServicesFragment : Fragment() {
-    private var _binding: FragmentServicesBinding? = null
+    private var _binding: FragmentCardsBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -22,10 +19,10 @@ class ServicesFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val servicesViewModel =
-            ViewModelProvider(this).get(ServicesViewModel::class.java)
+        val cardsViewModel =
+            ViewModelProvider(this).get(CardsViewModel::class.java)
 
-        _binding = FragmentServicesBinding.inflate(inflater, container, false)
+        _binding = FragmentCardsBinding.inflate(inflater, container, false)
 
         return binding.root
     }
