+++ app/src/main/java/com/example/bankodemia/ui/view/ContactAddedFragment.kt
@@ -5,11 +5,8 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.findNavController
 import com.example.bankodemia.R
-import com.example.bankodemia.UI.viewModel.CardsViewModel
-import com.example.bankodemia.databinding.FragmentCardsBinding
 import com.example.bankodemia.databinding.FragmentContactAddedBinding
 
 class ContactAddedFragment : Fragment() {
