+++ app/src/main/java/com/example/bankodemia/ui/view/HomeActivity.kt
@@ -1,4 +1,4 @@
-package com.example.bankodemia
+package com.example.bankodemia.ui.view
 
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -7,6 +7,7 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
+import com.example.bankodemia.R
 import com.example.bankodemia.databinding.ActivityHome2Binding
 
 class HomeActivity : AppCompatActivity() {
@@ -19,14 +20,16 @@ class HomeActivity : AppCompatActivity() {
         binding = ActivityHome2Binding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        setSupportActionBar(findViewById(R.id.homeToolBar))
+        supportActionBar?.hide()
         val navView: BottomNavigationView = binding.navView
-
+        navView.itemIconTintList = null
         val navController = findNavController(R.id.nav_host_fragment_activity_home2)
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.servicesFragment,R.id.cardsFragment,R.id.sendFragment
+                R.id.navigation_home, R.id.servicesFragment, R.id.cardsFragment, R.id.sendFragment
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
