+++ app/src/main/java/com/example/bankodemia/ui/view/SendFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bankodemia.UI.View
+package com.example.bankodemia.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -8,7 +8,7 @@ import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.findNavController
 import com.example.bankodemia.R
-import com.example.bankodemia.UI.viewModel.SendViewModel
+import com.example.bankodemia.ui.viewModel.SendViewModel
 import com.example.bankodemia.databinding.FragmentSendBinding
 
 class SendFragment : Fragment() {
