+++ app/src/main/java/com/example/bankodemia/ui/view/TransferDetailFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bankodemia.UI.View
+package com.example.bankodemia.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -12,9 +12,8 @@ import com.example.bankodemia.R
 import com.example.bankodemia.core.activateButton
 import com.example.bankodemia.core.validateField
 import com.example.bankodemia.databinding.FragmentTransferDetailBinding
-import com.example.bankodemia.ui.view.Fields
 
-class TransferDetailFragment : Frasgment(), Fields {
+class TransferDetailFragment : Fragment(), Fields {
     private var _binding: FragmentTransferDetailBinding? = null
     private val binding get() = _binding!!
