+++ app/src/main/res/drawable/ic_logomini.xml
@@ -2,6 +2,6 @@
     android:viewportWidth="38" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="#000000"
         android:pathData="M0,18.5l0,18.5 19,-0 19,-0 0,-18.5 0,-18.5 -19,-0 -19,-0 0,18.5zM26.8,5.6c8.2,5.3 8,18.5 -0.4,24 -3.5,2.3 -11.3,2.3 -14.8,-0 -10,-6.5 -8.1,-22.1 3,-25.6 3.6,-1.1 9.1,-0.3 12.2,1.6z" android:strokeColor="#00000000"/>
-    <path android:fillColor="#000000"
+    <path android:fillColor="#ffffff"
         android:pathData="M19,16.6c-1.4,4.8 -1.6,5 -4.4,2.9 -1.9,-1.4 -1.8,-1.5 0.5,-2.4 2.7,-1.1 0.6,-1.2 -3.1,-0.2 -2.4,0.7 -2.4,0.7 1.2,2.9 4.1,2.6 5.8,2.8 5.8,0.8 0,-0.8 0.4,-2.6 1,-4.1 0.9,-2.2 1.2,-2.4 3,-1.3 2.6,1.7 2.5,2.8 -0.2,2.9 -1.4,0.1 -1.7,0.3 -0.7,0.6 0.8,0.2 2.6,-0 3.8,-0.5 2.2,-0.8 2.1,-0.9 -1.7,-3.4l-4,-2.7 -1.2,4.5z" android:strokeColor="#00000000"/>
 </vector>
