+++ app/src/main/res/drawable/nav_view_item_textcolot.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/principal_blue" android:state_checked="true" />
+    <item android:color="@color/white" />
+</selector>
