+++ app/src/main/res/layout/fragment_home_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".UI.View.HomeDetailFragment">
+    tools:context=".ui.view.HomeDetailFragment">
 
     <include
         android:id="@+id/homeDetailTitleBar"
