+++ app/src/main/res/layout/fragment_send.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".UI.View.SendFragment">
+    tools:context=".ui.view.SendFragment">
 
     <ImageView
         android:id="@+id/sendIvLogo"
