+++ app/src/main/res/layout/fragment_services.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".UI.View.ServicesFragment">
+    tools:context=".ui.view.ServicesFragment">
 
     <include
         android:id="@+id/servicesTemplateTitleBar"
