+++ app/src/main/res/layout/fragment_transfer_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".UI.View.TransferDetailFragment">
+    tools:context=".ui.view.TransferDetailFragment">
 
     <include
         android:id="@+id/detailTemplateTopLogoSingle"
